week3/pyproject.toml
[tool.poetry]
name = "week3"
version = "0.1.0"
description = "Hasan infra week3"
authors = ["treblenalto <taehee107275@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9"
black = "^23.12.1"
boto3 = "^1.34.11"
pandas = "^2.1.4"
redis = "^5.0.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
